feat(ui-plan): replace PlanScreen with create-only flow
